requirements-search.txt
streamlit==1.56.0
streamlit-authenticator==0.4.2
pinecone==8.1.2
openai==2.32.0
python-dotenv==1.2.2
pyyaml==6.0.3
